pyproject.toml
[project]
name = "prefect"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.0",
    "fastmcp>=3.2.4",
    "faststream[rabbit]>=0.6.7",
    "logfire>=4.32.1",
    "pyaml>=26.2.1",
    "pydantic-ai>=1.84.1",
    "uvicorn>=0.44.0",
    "watchdog>=6.0.0",
    "croniter>=6.0.0",
    "rich>=15.0.0",
    "typer>=0.24.1",
    "jinja2>=3.1.6",
    "spectree>=2.0.1",
    "pydantic-evals[logfire]>=1.84.1",
    "better-exceptions>=0.3.3",
]
